Relocated get_species_by_label from main to common
