+++ t3/common.py
@@ -4,6 +4,9 @@ t3 common module
 
 import os
 import string
+from typing import Optional
+
+from rmgpy.species import Species
 
 
 VERSION = '0.1.0'
@@ -13,6 +16,31 @@ DATA_BASE_PATH = os.path.join(t3_path, 'tests', 'data')
 EXAMPLES_BASE_PATH = os.path.join(t3_path, 'examples')
 PROJECTS_BASE_PATH = os.path.join(t3_path, 'Projects')
 VALID_CHARS = "-_=.,%s%s" % (string.ascii_letters, string.digits)
+
+
+def get_species_by_label(label: str,
+                         species_list: list,
+                         ) -> Optional[Species]:
+    """
+    Get a species from a list of species by its label.
+
+    Args:
+        label (str): A species label.
+        species_list (list): Entries are RMG Species objects.
+
+    Returns:
+        Optional[Species]: The corresponding species from the species_list.
+                           Returns ``None`` if no species was found.
+    """
+    for species in species_list:
+        if species.label == label or species.to_chemkin() == label:
+            return species
+    if '(' in label and ')' in label:
+        # try by the RMG species index
+        for species in species_list:
+            if species.index == int(label.split('(')[-1].split(')')[0]):
+                return species
+    return None
 
 
 def dict_to_str(dictionary: dict,
